function_app/requirements.txt
azure-functions==1.20.0
fastapi==0.116.1
uvicorn==0.35.0
opencv-python==4.12.0.88
pytesseract==0.3.13
pillow==11.3.0
python-multipart==0.0.20